+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/task/dao/TaskDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Transactional
 @Repository
-public interface TaskDao extends CrudRepository<Task, Integer> {
+public interface TaskDao extends CrudRepository<Task, Long> {
     List<Task> findByDuration(int duration);
 
 }
