+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/task/dao/TaskFinancialDetailsDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Transactional
 @Repository
-public interface TaskFinancialDetailsDao extends CrudRepository<TaskFinancialDetails, Integer> {
+public interface TaskFinancialDetailsDao extends CrudRepository<TaskFinancialDetails, Long> {
     List<TaskFinancialDetails> findByPaid(boolean paid);
 
 }
