+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/tasklist/TaskList.java
@@ -6,7 +6,7 @@ import javax.validation.constraints.NotNull;
 @Entity
 @Table(name = "TASKLISTS")
 public class TaskList {
-    private int id;
+    private long id;
     private String listName;
     private String description;
 
@@ -22,7 +22,7 @@ public class TaskList {
     @Id
     @GeneratedValue
     @Column(name = "ID", unique = true)
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -41,7 +41,7 @@ public class TaskList {
         this.description = description;
     }
 
-    private void setId(int id) {
+    private void setId(long id) {
         this.id = id;
     }
