+++ kodilla-hibernate/src/main/java/com/kodilla/hibernate/tasklist/dao/TaskListDao.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 @Transactional
 @Repository
-public interface TaskListDao extends CrudRepository<TaskList, Integer> {
+public interface TaskListDao extends CrudRepository<TaskList, Long> {
     List<TaskList> findByListName(String listName);
 }
