+++ kodilla-hibernate/src/test/java/com/kodilla/hibernate/task/dao/TaskDaoTestSuite.java
@@ -29,7 +29,7 @@ public class TaskDaoTestSuite {
         taskDao.save(task);
 
         //Then
-        final int id = task.getId();
+        final long id = task.getId();
         final Task readTask = taskDao.findOne(id);
         Assert.assertEquals(id, readTask.getId());
 
@@ -51,7 +51,7 @@ public class TaskDaoTestSuite {
         Assert.assertEquals(1, readTasks.size());
 
         //CleanUp
-        final int id = readTasks.get(0).getId();
+        final long id = readTasks.get(0).getId();
         taskDao.delete(id);
     }
 
@@ -63,7 +63,7 @@ public class TaskDaoTestSuite {
 
         //When
         taskDao.save(task);
-        final int id = task.getId();
+        final long id = task.getId();
 
         //Then
         Assert.assertNotEquals(0, id);
